Replace NaN-values with zero
SNAPHU errors if the input interferogram contains NaN or infinity
values, but NaN is commonly used as a fill value to represent invalid
samples, which SNAPHU should mask out for purposes of unwrapping.

It's convenient for the Python wrapper to replace NaN values with zeros,
since SNAPHU treats zero-magnitude samples as invalid (masked out).