Add option to regrow connected components in tiled mode
Add an option to re-run SNAPHU after the initial unwrapping step to
"regrow" the connected components. This is useful in tiled unwrapping
mode to compute a new set of connected components that are not delimited
by the tile boundaries.

This commit also removes the `SnaphuConfig` class that was a bit
over-engineered and wasn't easily extendable to support some of the more
esoteric SNAPHU configurations/modes.