Add an option (enabled by default) to regrow connected components in tiled mode (#21)
* Add option to regrow connected components in tiled mode

Add an option to re-run SNAPHU after the initial unwrapping step to
"regrow" the connected components. This is useful in tiled unwrapping
mode to compute a new set of connected components that are not delimited
by the tile boundaries.

This commit also removes the `SnaphuConfig` class that was a bit
over-engineered and wasn't easily extendable to support some of the more
esoteric SNAPHU configurations/modes.

* Ensure config files are uniquely named in scratch directory

Currently, we're careful to ensure that intermediate files created in
the scratch directory are created safely using `mkstemp()`, since a user
might run multiple unwrapping processes in parallel with the same
scratch directory.

However, we weren't as careful about naming the config file(s) -- their
names were simply hard-coded. This is now resolved by using `mkstemp()`
to create the config files as well.