Remove local version identifiers
By default, setuptools_scm applies "local version identifiers" to
non-tag versions [^1]. In practice, this means appending the short git
hash as a suffix to the version string.

However, TestPyPI doesn't allow publishing distributions that include
local version labels [^2][^3][^4]. This is intended to prevent
accidental upload of non-release versions, but it makes it a bit more
complicated to test our CD pipeline.

We could try to work around this by spoofing the version using
`SETUPTOOLS_SCM_PRETEND_VERSION`[^5]. Alternatively, we could simply
disable local version identifiers in the setuptools_scm configuration.

Including the git hash in the version string of non-tag builds seems
very useful, but handling the version spoofing in CI seems much more
complicated than I'm willing to tolerate.

So for now we'll forego the local version labels, but in the future it'd
be nice to reenable them and adopt a better workaround.

[^1]: https://packaging.python.org/en/latest/specifications/version-specifiers/#local-version-identifiers
[^2]: https://github.com/pypa/setuptools_scm/issues/322
[^3]: https://github.com/pypa/setuptools_scm/issues/455
[^4]: https://github.com/pypa/setuptools_scm/issues/478
[^5]: https://setuptools-scm.readthedocs.io/en/latest/overrides/