Add GHA job to publish to PyPI
Add a new CI job that triggers whenever a new tag is pushed. It collects
the wheels and sdist generated by the `build-wheels` and `build-sdist`
jobs and uploads them to PyPI. This job requires manual approval [^1].

Also, consolidate `cd.yml` into `ci.yml`. It's convenient for all of the
CI jobs to be defined in the same YAML file in order to easily introduce
interdependencies between jobs. In this case, we want the
`create-release` job to depend on all of the CI testing jobs (so that a
release will not be created if any tests fail). We also want the
`publish-to-pypi` job to depend on the `create-release` job (so that we
don't upload the distribution to PyPI if a release was not created for
some reason).

[^1]:
https://docs.github.com/en/actions/deployment/targeting-different-environments/using-environments-for-deployment#deployment-protection-rules