Improve Pylance support for `Raster`
Pylance seemed to be failing to find the definition of
`snaphu.io.Raster` (presumably due to the heterodox way in which it's
exported), which manifested in VSCode tooltips implicitly treating the
type as equivalent to `Any`.

This update tries to update the way that the Raster symbol is exported
to improve compatibility with Pylance's behavior while still preserving
the same runtime behavior. In particular:

1. If the `rasterio` package is missing, importing `snaphu.io` should
   not fail. `Raster` should not be a member of `snaphu.io` in this
   case. Attempting to import `snaphu.io.Raster` should raise an
   `ImportError`.
2. If the `rasterio` package is available, `snaphu.io.Raster` should be
   a valid type and `'Raster'` should be a member of `dir(snaphu.io)`.