Ensure memory-mapped files are safely closed
* Add `io.MMapArray`, an array_like type that uses a memory-mapped file
  buffer for storage
* Replace usage of `numpy.memmap` with `io.MMapArray`

`numpy.memmap` does not have an API that guarantees that the underlying
file is closed. This could lead to a large number of file descriptors
staying open for much longer than necessary.

>Currently there is no API to close the underlying mmap. It is tricky to
>ensure the resource is actually closed, since it may be shared between
>different memmap instances.

https://numpy.org/doc/stable/reference/generated/numpy.memmap.html

In practice, this seems to have been the cause of errors encountered in
testing due to too many open files when unwrapping multiple
interferograms in parallel.

```
OSError: [Errno 24] Too many open files
```